chronobot::chrono

Module dice

This module govorns dice-rolls and their behaviour, including deserializing the rules from a csv.

The rules are read from <bin-dir>/tables/csv/dice.csv, and must be in the following format:

| Name (String) | Number of Sides (Positive Integer) | Associated Color (String) | Results |

Here’s an example row:

Advantage,6,Blue,//S/SO/OO/O

This will be deserialized using std::str::FromStr and From<char>, into a list of DiceRollTableItem.

An entry for “Results” is a sequence of outcomes delimited by a / character, with each character representing a score: | Character | Meaning | | | Blank | | B | Blank | | S | Success | | F | Failure | | O | Opportunity | | T | Threat | | * | Triumph | | X | Despair |

NOTE: “blank” can be denoted either by a lack of any characters, or by a single “B” character

The number of items delimited by the / character MUST match the value provided for the “Number of Sides” field.

For example, the current “advantage” roll is defined as

blank | blank | green | green blue | blue blue | blue

So the corresponding csv entry would be //S/SO/OO/O.

After all entries have been read from the CSV, they are collected into ROLL_TABLES

For convenience, the RollAxes can be added together to evaluate the total outcome.

let a = RollAxes::PassFail{ outcome: true, outstanding: true };
let b = RollAxes::PassFail{ outcome: false, outstanding: false };

// DiceRollResult {
//     passfail:  0,
//     blank:     0,
//     luck:      0,
//     triumph:   1,
//     despair:   0,
// }
let result: DiceRollResult = &a + &b;

Structs

DiceRollResult

DiceRollTableItem

IoError

RollAxesVec 🔒

Enums

RollAxes

RollError

Statics

ROLL_TABLES

Functions

read_dice_table 🔒

roll_dice